[gallery] refactored getStyleByLayout in layoutHelper to gallery/presets
