+++ packages/gallery/src/components/gallery/presets/alternateGallery.js
@@ -5,21 +5,43 @@ import LAYOUTS from '../../../common/constants/layout';
 
 export const fixedStyles = {
   galleryLayout: LAYOUTS.ALTERNATE,
+  //params from layoutHelper
+  sampleSize: 100,
+  isVertical: true,
+  gallerySize: 86,
+  minItemSize: 50,
+  groupSize: 3,
+  chooseBestGroup: true,
+  groupTypes: '1,2h,2v,3t,3b,3l,3r,3v,3h',
+  rotatingGroupTypes: '1,3l,1,3r',
+  cubeImages: true,
+  cubeType: 'fill',
+  smartCrop: false,
+  collageDensity: 0.48,
+  galleryMargin: 0,
+  floatingImages: 0,
+  cubeRatio: 1,
+  fixedColumns: 1,
+  groupsPerStrip: 0,
+  oneRow: false,
+  placeGroupsLtr: false,
+  at: 1538490828979,
+  rotatingCropRatios: '',
 }
 export default class alternateGallery extends React.Component {
 
-    render() {
+  render() {
 
-        
 
-        return (
-            <ProGallery
-                {...this.props}
-                styles={{
-                    ...this.props.styles,
-                    ...fixedStyles
-                }}
-            />
-        );
-    }
+
+    return (
+      <ProGallery
+        {...this.props}
+        styles={{
+          ...this.props.styles,
+          ...fixedStyles
+        }}
+      />
+    );
+  }
 }
