+++ packages/gallery/src/components/gallery/presets/slideshowGallery.js
@@ -5,7 +5,31 @@ import LAYOUTS from '../../../common/constants/layout';
 export const fixedStyles = {
   galleryLayout: LAYOUTS.SLIDESHOW,
   enableInfiniteScroll: true,
-  allowHover: false
+  allowHover: false,
+  //params from layoutHelper
+  showArrows: true,
+  cubeImages: true,
+  smartCrop: false,
+  cubeRatio: '100%/100%',
+  isVertical: false,
+  gallerySize: 550,
+  galleryType: 'Strips',
+  groupSize: 1,
+  groupTypes: '1',
+  fixedColumns: 1,
+  oneRow: true,
+  hasThumbnails: false,
+  enableScroll: true,
+  scrollSnap: true,
+  isGrid: false,
+  isColumns: false,
+  isMasonry: false,
+  isSlider: false,
+  isSlideshow: true,
+  cropOnlyFill: false,
+  floatingImages: 0,
+  galleryMargin: 0,
+  imageMargin: 0,
 }
 export default class SlideshowGallery extends React.Component {
   render() {
